Get `all_open_btc_dai_orders` query to queries module
