+++ cnd/src/storage/db/queries.rs
@@ -4,7 +4,14 @@
 //! the codebase, we expose queries that compose together diesel's primitives.
 
 use crate::{
-    storage::{db::schema::swap_contexts, NoSwapExists, SwapContext, Text},
+    asset,
+    storage::{
+        db::{
+            schema::{btc_dai_orders, orders, swap_contexts},
+            wrapper_types::Satoshis,
+        },
+        BtcDaiOrder, NoSwapExists, Order, SwapContext, Text,
+    },
     LocalSwapId,
 };
 use anyhow::Result;
@@ -24,4 +31,14 @@ pub fn get_all_swap_contexts(conn: &SqliteConnection) -> Result<Vec<SwapContext>
     let contexts = swap_contexts::table.load::<SwapContext>(conn)?;
 
     Ok(contexts)
+}
+
+pub fn all_open_btc_dai_orders(conn: &SqliteConnection) -> Result<Vec<(Order, BtcDaiOrder)>> {
+    let orders = orders::table
+        .inner_join(btc_dai_orders::table)
+        .filter(btc_dai_orders::open.ne(Text::<Satoshis>(asset::Bitcoin::ZERO.into())))
+        .or_filter(btc_dai_orders::settling.ne(Text::<Satoshis>(asset::Bitcoin::ZERO.into())))
+        .load::<(Order, BtcDaiOrder)>(conn)?;
+
+    Ok(orders)
 }
